dev/architercture/lab1/backend/requirements.txt
djangorestframework
markdown
django-filter
psycopg[binary,pool]